eseva_portal/requirements.txt
Flask>=3.0.0
pymongo>=4.6.1
python-dotenv>=1.0.1
opencv-python-headless>=4.10.0.84
pytesseract>=0.3.10
google-api-python-client>=2.111.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
requests>=2.31.0
Werkzeug>=3.0.1
Flask-PyMongo>=2.3.0
flask-cors>=4.0.0
numpy>=2.1.0
